refactor(services): another one bites the dust
