+++ Cargo.toml
@@ -37,6 +37,7 @@ enum_meta = "=0.7.0"
 itertools = "=0.13.0"
 markdown = "=1.0.0-alpha.18"
 nanoid = "=0.4.0"
+regex = "=1.10.5"
 rust_decimal = "=1.35.0"
 rust_decimal_macros = "=1.35.0"
 schematic = { version = "=0.16.6", features = [
@@ -63,6 +64,7 @@ sea-orm = { version = "=1.0.0", features = [
     "with-uuid",
 ], default-features = false }
 sea-orm-migration = "=1.0.0"
+sea-query = "=0.31.0"
 serde = { version = "=1.0.204", features = ["derive"] }
 serde_json = "=1.0.120"
 serde_with = { version = "=3.9.0", features = ["chrono_0_4"] }
