+++ apps/backend/Cargo.toml
@@ -36,8 +36,7 @@ models = { path = "../../crates/models" }
 nanoid = { workspace = true }
 openidconnect = "=3.5.0"
 providers = { path = "../../crates/providers" }
-radarr-api-rs = "=3.0.1"
-regex = "=1.10.5"
+regex = { workspace = true }
 reqwest = { workspace = true }
 resolvers = { path = "../../crates/resolvers" }
 rust_decimal = { workspace = true }
@@ -45,14 +44,13 @@ rust_decimal_macros = { workspace = true }
 schematic = { workspace = true }
 sea-orm = { workspace = true }
 sea-orm-migration = { workspace = true }
-sea-query = "=0.31.0"
+sea-query = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_with = { workspace = true }
 serde-xml-rs = "=0.6.0"
 services = { path = "../../crates/services" }
 slug = "=0.1.5"
-sonarr-api-rs = "=3.0.0"
 tokio = { workspace = true }
 tower = { version = "=0.4.13", features = ["buffer"] }
 tower-http = { version = "=0.5.2", features = ["catch-panic", "cors", "trace"] }
