+++ crates/services/integration/Cargo.toml
@@ -4,3 +4,24 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = { workspace = true }
+application_utils = { path = "../../utils/application" }
+async-graphql = { workspace = true }
+enums = { path = "../../enums" }
+database_models = { path = "../../models/database" }
+database_utils = { path = "../../utils/database" }
+media_models = { path = "../../models/media" }
+providers = { path = "../../providers" }
+radarr-api-rs = "=3.0.1"
+regex = { workspace = true }
+reqwest = { workspace = true }
+rust_decimal = { workspace = true }
+rust_decimal_macros = { workspace = true }
+sea-orm = { workspace = true }
+sea-query = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+sonarr-api-rs = "=3.0.0"
+specific_models = { path = "../../models/specific" }
+tracing = { workspace = true }
+traits = { path = "../../traits" }
